A functional tweak to favourite button
Favourite button is used to be used only to set a notification for the corresponding session, so you couldn't be able to favourite a passed session.
But from now on, you can favourite every session even if it's passed.
Notification will be set only for the future sessions if they are favourited.